moved container utils function
