+++ dags/container_operators.py
@@ -1,8 +1,6 @@
-import json
 import logging
 from pathlib import Path
 from tempfile import mkdtemp
-from typing import List
 
 import yaml
 
@@ -10,6 +8,9 @@ from airflow.models import DAG
 from airflow.operators.bash import BashOperator
 
 from sciencebeam_airflow.utils.container import (
+    _get_prefer_preemptible_json,
+    _get_select_preemptible_json,
+    _get_helm_select_preemptible_values,
     get_helm_delete_command
 )
 
@@ -17,67 +18,6 @@ from sciencebeam_dag_utils import add_dag_macro
 
 
 LOGGER = logging.getLogger(__name__)
-
-
-def _get_preemptible_affinity():
-    return {
-        "nodeAffinity": {
-            "preferredDuringSchedulingIgnoredDuringExecution": [{
-                "weight": 1,
-                "preference": {
-                    "matchExpressions": [{
-                        "key": "cloud.google.com/gke-preemptible",
-                        "operator": "In",
-                        "values": ["true"]
-                    }]
-                }
-            }]
-        }
-    }
-
-
-def _get_preemptible_toleration():
-    return {
-        "key": "cloud.google.com/gke-preemptible",
-        "operator": "Equal",
-        "value": "true",
-        "effect": "NoSchedule"
-    }
-
-
-def _get_prefer_preemptible_spec():
-    return {
-        "affinity": _get_preemptible_affinity(),
-        "tolerations": [_get_preemptible_toleration()]
-    }
-
-
-def _get_select_preemptible_spec():
-    return {
-        "nodeSelector": {
-            "cloud.google.com/gke-preemptible": "true"
-        },
-        "tolerations": [_get_preemptible_toleration()]
-    }
-
-
-def _get_prefer_preemptible_json():
-    return json.dumps({
-        "spec": _get_prefer_preemptible_spec()
-    })
-
-
-def _get_select_preemptible_json():
-    return json.dumps({
-        "spec": _get_select_preemptible_spec()
-    })
-
-
-def _get_helm_preemptible_values(child_chart_names: List[str] = None) -> dict:
-    values = _get_select_preemptible_spec().copy()
-    for child_chart_name in (child_chart_names or []):
-        values[child_chart_name] = _get_select_preemptible_spec()
-    return values
 
 
 def generate_run_name(name: str, suffix: str = '', other_suffix: str = '') -> str:
@@ -172,7 +112,7 @@ class HelmDeployOperator(BashOperator):
             child_chart_names = []
             if self.get_child_chart_names:
                 child_chart_names = self.get_child_chart_names(dag_run=dag_run)
-            values_file.write_text(yaml.safe_dump(_get_helm_preemptible_values(
+            values_file.write_text(yaml.safe_dump(_get_helm_select_preemptible_values(
                 child_chart_names
             )))
             LOGGER.info('helm values (%s): %s', values_file, values_file.read_text())
